refactor: Move member.h to support/
